+++ src/main/java/kitchenpos/order/domain/OrderRepository.java
@@ -1,5 +1,6 @@
 package kitchenpos.order.domain;
 
+import kitchenpos.table.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
