+++ src/main/java/kitchenpos/table/dto/TableGroupResponse.java
@@ -1,6 +1,6 @@
-package kitchenpos.order.dto;
+package kitchenpos.table.dto;
 
-import kitchenpos.order.domain.TableGroup;
+import kitchenpos.table.domain.TableGroup;
 
 import java.time.LocalDateTime;
 import java.util.List;
