+++ src/test/java/kitchenpos/testfixture/CommonTestFixture.java
@@ -9,8 +9,10 @@ import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.order.domain.*;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.order.dto.OrderRequest;
-import kitchenpos.order.dto.OrderTableRequest;
-import kitchenpos.order.dto.TableGroupRequest;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
+import kitchenpos.table.dto.OrderTableRequest;
+import kitchenpos.table.dto.TableGroupRequest;
 import kitchenpos.product.domain.*;
 import kitchenpos.product.dto.ProductRequest;
